Exploit lexical closures to allow byte-compilation of wrapped functions.
A number of functions are wrapped in a thin anonymous function before being
stored in the data structures. The wrappers were constructed as quoted lambda
forms, so were not byte-compiled. Lexical closures allow the
wrapper-constructing functions to use unquoted lambdas, enabling
byte-compilation of the wrapped functions.

We fall back to old-style function wrapping if lexical binding unsupported.